+++ persistent/Database/Persist/Sql/Types/Internal.hs
@@ -81,6 +81,11 @@ data SqlBackend = SqlBackend
     , connRDBMS :: Text
     , connLimitOffset :: (Int,Int) -> Bool -> Text -> Text
     , connLogFunc :: LogFunc
+    , connMaxParams :: Maybe Int
+    -- ^ Some databases (probably only Sqlite) have a limit on how
+    -- many question-mark parameters may be used in a statement
+    --
+    -- @since 2.6.1
     }
     deriving Typeable
 instance HasPersistBackend SqlBackend where